+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
+## [Unreleased]
+
+### Fixed
+
+- extract trivy-scan: model error on empty Results and Vulnerabilities list
+
+
 ## [0.6.0] 2025-02-04
 
 ### Added